Refactor Petrinaut core handles
