+++ libs/@hashintel/petrinaut-core/src/handle/index.ts
@@ -1,0 +1,13 @@
+export {
+  createJsonDocHandle,
+  type CreateJsonDocHandleOptions,
+} from "./json-doc-handle";
+export type {
+  DocChangeEvent,
+  DocHandleState,
+  DocumentId,
+  HistoryEntry,
+  PetrinautDocHandle,
+  PetrinautHistory,
+  PetrinautPatch,
+} from "./types";
