+++ libs/@hashintel/petrinaut-core/src/handle/json-doc-handle/create-json-doc-handle.ts
@@ -5,96 +5,21 @@ import {
   produceWithPatches,
 } from "immer";
 
-import type { SDCPN } from "./types/sdcpn";
-import type { PetrinautHandleCapabilities } from "./extensions";
+import { createReadableStore } from "../../store";
+
+import type { PetrinautHandleCapabilities } from "../../extensions";
+import type { SDCPN } from "../../types/sdcpn";
+import type {
+  DocChangeEvent,
+  DocHandleState,
+  DocumentId,
+  HistoryEntry,
+  PetrinautDocHandle,
+  PetrinautHistory,
+  PetrinautPatch,
+} from "../types";
 
 enablePatches();
-
-export type DocumentId = string;
-
-export type DocHandleState = "loading" | "ready" | "deleted" | "unavailable";
-
-export type PetrinautPatch = {
-  op: "add" | "remove" | "replace";
-  path: (string | number)[];
-  value?: unknown;
-};
-
-export type DocChangeEvent = {
-  next: SDCPN;
-  patches?: PetrinautPatch[];
-  source?: "local" | "remote";
-};
-
-export type ReadableStore<T> = {
-  get(): T;
-  subscribe(listener: (value: T) => void): () => void;
-};
-
-export type HistoryEntry = {
-  timestamp: string;
-};
-
-export interface PetrinautHistory {
-  /** Apply the most recent inverse patches. Returns true if anything was undone. */
-  undo(): boolean;
-  /** Re-apply the most recently undone patches. Returns true if anything was redone. */
-  redo(): boolean;
-  /** Jump to an arbitrary point in history. Returns true if the index changed. */
-  goToIndex(index: number): boolean;
-  /** Drop the entire history. */
-  clear(): void;
-
-  readonly canUndo: ReadableStore<boolean>;
-  readonly canRedo: ReadableStore<boolean>;
-
-  /**
-   * Ordered timestamps of the history checkpoints. Index 0 is the initial
-   * state; index N is the state after the Nth mutation.
-   */
-  readonly entries: ReadableStore<readonly HistoryEntry[]>;
-
-  /** Position of the current state within {@link entries}. */
-  readonly currentIndex: ReadableStore<number>;
-}
-
-export interface PetrinautDocHandle {
-  readonly id: DocumentId;
-  readonly capabilities?: PetrinautHandleCapabilities;
-  readonly state: ReadableStore<DocHandleState>;
-  whenReady(): Promise<void>;
-  doc(): SDCPN | undefined;
-  change(fn: (draft: SDCPN) => void): void;
-  subscribe(listener: (event: DocChangeEvent) => void): () => void;
-  /**
-   * Optional. Present on handles that track local history (Immer-backed,
-   * Automerge-backed, …). Read-only mirror handles omit it.
-   */
-  readonly history?: PetrinautHistory;
-}
-
-function createReadableStore<T>(initial: T): ReadableStore<T> & {
-  set(next: T): void;
-} {
-  let current = initial;
-  const listeners = new Set<(value: T) => void>();
-  return {
-    get: () => current,
-    subscribe(listener) {
-      listeners.add(listener);
-      return () => listeners.delete(listener);
-    },
-    set(next) {
-      if (Object.is(next, current)) {
-        return;
-      }
-      current = next;
-      for (const listener of listeners) {
-        listener(current);
-      }
-    },
-  };
-}
 
 function fromImmerPatch(patch: ImmerPatch): PetrinautPatch {
   return {
@@ -130,6 +55,17 @@ export type CreateJsonDocHandleOptions = {
   historyLimit?: number;
 };
 
+/**
+ * Create an in-memory JSON-backed Petrinaut document handle.
+ *
+ * Petrinaut is built around the `PetrinautDocHandle` contract so hosts can
+ * provide different document backends, such as local JSON state, collaborative
+ * documents, or read-only mirrors. This implementation is the small default
+ * handle used by tests, stories, examples, and simple local editing flows. It
+ * stores the current SDCPN snapshot in memory, publishes local change events
+ * with Immer patches, exposes optional undo/redo history, and honors handle
+ * capabilities such as `readonly`.
+ */
 export function createJsonDocHandle(
   opts: CreateJsonDocHandleOptions,
 ): PetrinautDocHandle {
