+++ libs/@hashintel/petrinaut-core/src/handle/json-doc-handle/index.ts
@@ -1,0 +1,4 @@
+export {
+  createJsonDocHandle,
+  type CreateJsonDocHandleOptions,
+} from "./create-json-doc-handle";
